feat(people): blueprint endpoints + restore alembic; fix /people/find response
